requirements.txt
tensorflow==2.15.0
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
pillow==10.1.0
editdistance==0.8.1
scikit-learn==1.3.2
matplotlib==3.8.2 